apps/ads-manager-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
APScheduler==3.10.4
asyncpg==0.29.0
python-multipart==0.0.12
